pyproject.toml
[project]
name = "llm-fingerprint-daily"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "apscheduler>=3.10",
    "typer>=0.12",
    "jinja2>=3.1",
    "scikit-learn>=1.4",
    "scipy>=1.12",
    "numpy>=1.26",
    "pyyaml>=6.0",
]

[project.scripts]
fingerprint = "src.main:app"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "pytest-cov"]

[tool.setuptools.packages.find]
include = ["src*"]
